Move cpp20_rwc_syncs into compile-fail
